requirements.txt
opencc_python_reimplemented==0.1.7
